pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stelline"
version = "2.0.0"
description = "Session intelligence for AI agents — crafts authentic memories from conversation logs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Haseeb Khalid", email = "haseeb@mahamedia.us"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agents", "memory", "session", "intelligence", "rag"]

dependencies = [
    "click>=8.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
stelline = "stelline.cli:cli"

[project.urls]
Homepage = "https://github.com/HaseebKhalid1507/stelline"
Repository = "https://github.com/HaseebKhalid1507/stelline"
Issues = "https://github.com/HaseebKhalid1507/stelline/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["stelline*"]
